requirements.txt
beautifulsoup4==4.6.0
certifi==2018.8.24
chardet==3.0.4
Django==2.0.8
django-modelcluster==4.2
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.8.2
docopt==0.4.0
draftjs-exporter==2.1.2
html5lib==1.0.1
idna==2.7
mailchimp==2.0.9
Pillow==5.2.0
pytz==2018.5
requests==2.19.1
six==1.11.0
Unidecode==1.0.22
urllib3==1.23
wagtail==2.2.2
webencodings==0.5.1
Willow==1.1
